Fix bad migration
I accidentally moved the controller customizations fro SequenceBlocks to
Sequences. Fixed that here.
